Fix the passing variables to the simulation
Next step is getting the simulation fixed and observe the values from
simulation, verify what variables is often ploted in the c++ Program.

+++ Project/src/algorithm/Simulation.java
@@ -10,7 +10,7 @@ import static java.lang.Math.*;
 public class Simulation {
 	
 	/** Guarda o tempo de simulação */
-	public float tempo;
+	public double tempo;
 	/** Identifica a simulação (não usada) */
 	public int idSimulacao;
 	/** Identifica a simulação (não usada) */
